# src/infrastructure/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Jummall API",
    "version": "1.0.0",
    "description": "API documentation for admin, seller, and buyer domains"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login any user role",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Account is not verified"
          },
          "429": {
            "description": "Too many login attempts"
          }
        }
      }
    },
    "/auth/verify-email": {
      "post": {
        "tags": ["auth"],
        "summary": "Verify a user's email with a code",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired verification code"
          },
          "404": {
            "description": "Verification request not found"
          },
          "409": {
            "description": "Account already verified"
          }
        }
      }
    },
    "/auth/resend-verification-email": {
      "post": {
        "tags": ["auth"],
        "summary": "Resend a verification email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification email resent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResendVerificationEmailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Verification request not found"
          },
          "409": {
            "description": "Account already verified"
          },
          "429": {
            "description": "Resend requested too soon"
          }
        }
      }
    },
    "/buyers/register": {
      "post": {
        "tags": ["buyer"],
        "summary": "Register a buyer account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Buyer registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "409": {
            "description": "Buyer credentials already exist"
          }
        }
      }
    },
    "/sellers/register": {
      "post": {
        "tags": ["seller"],
        "summary": "Register a seller account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellerRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Seller registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "409": {
            "description": "Seller credentials already exist"
          }
        }
      }
    },
    "/sellers/kyc": {
      "put": {
        "tags": ["seller"],
        "summary": "Save seller KYC draft",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveSellerKycDraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seller KYC draft saved successfully"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "KYC cannot be updated in the current state"
          }
        }
      }
    },
    "/sellers/kyc/documents": {
      "post": {
        "tags": ["seller"],
        "summary": "Upload a seller KYC document",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadSellerKycDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Seller KYC document uploaded successfully"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "Document cannot be updated in the current KYC state"
          }
        }
      }
    },
    "/sellers/kyc/submit": {
      "post": {
        "tags": ["seller"],
        "summary": "Submit seller KYC for review",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitSellerKycRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Seller KYC submitted successfully"
          },
          "400": {
            "description": "KYC submission is incomplete"
          },
          "401": {
            "description": "Authentication required"
          },
          "409": {
            "description": "KYC cannot be submitted in the current state"
          }
        }
      }
    }
  },
  "tags": [
    { "name": "auth" },
    { "name": "admin" },
    { "name": "seller" },
    { "name": "buyer" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "Password123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful."
          },
          "data": {
            "type": "object",
            "properties": {
              "access_token": {
                "type": "string"
              },
              "user": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "first_name": {
                    "type": "string",
                    "nullable": true
                  },
                  "last_name": {
                    "type": "string",
                    "nullable": true
                  },
                  "username": {
                    "type": "string",
                    "nullable": true
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "phone": {
                    "type": "string",
                    "nullable": true
                  },
                  "role": {
                    "type": "string",
                    "example": "buyer"
                  },
                  "account_status": {
                    "type": "string",
                    "example": "verified"
                  },
                  "created_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "updated_at": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        }
      },
      "VerifyEmailRequest": {
        "type": "object",
        "required": ["email", "code"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "code": {
            "type": "string",
            "example": "123456"
          }
        }
      },
      "VerifyEmailResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Email verified successfully."
          },
          "data": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "format": "email"
              },
              "account_status": {
                "type": "string",
                "example": "verified"
              }
            }
          }
        }
      },
      "ResendVerificationEmailRequest": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          }
        }
      },
      "ResendVerificationEmailResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Verification email resent successfully."
          },
          "data": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "format": "email"
              }
            }
          }
        }
      },
      "BuyerRegisterRequest": {
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "username",
          "email",
          "password",
          "phone",
          "password_confirmation"
        ],
        "properties": {
          "first_name": {
            "type": "string",
            "example": "John"
          },
          "last_name": {
            "type": "string",
            "example": "Doe"
          },
          "username": {
            "type": "string",
            "example": "john.doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "Password123"
          },
          "phone": {
            "type": "string",
            "example": "+2348012345678"
          },
          "password_confirmation": {
            "type": "string",
            "format": "password",
            "example": "Password123"
          }
        }
      },
      "BuyerRegisterResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Buyer registered successfully."
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "first_name": {
                "type": "string"
              },
              "last_name": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "phone": {
                "type": "string"
              },
              "role": {
                "type": "string",
                "example": "buyer"
              },
              "account_status": {
                "type": "string",
                "example": "not_verified"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
      "SellerRegisterRequest": {
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "username",
          "email",
          "phone_number",
          "password",
          "confirm_password",
          "account_type"
        ],
        "properties": {
          "first_name": {
            "type": "string",
            "example": "Jane"
          },
          "last_name": {
            "type": "string",
            "example": "Doe"
          },
          "username": {
            "type": "string",
            "example": "jane.doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane@example.com"
          },
          "phone_number": {
            "type": "string",
            "example": "+2348012345678"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "Password123"
          },
          "confirm_password": {
            "type": "string",
            "format": "password",
            "example": "Password123"
          },
          "account_type": {
            "type": "string",
            "enum": ["individual", "business"],
            "example": "individual"
          }
        }
      },
      "SellerRegisterResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Seller registered successfully."
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "first_name": {
                "type": "string"
              },
              "last_name": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "phone_number": {
                "type": "string"
              },
              "role": {
                "type": "string",
                "example": "seller"
              },
              "account_status": {
                "type": "string",
                "example": "not_verified"
              },
              "account_type": {
                "type": "string",
                "example": "individual"
              },
              "kyc_status": {
                "type": "string",
                "example": "not_started"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
      "SaveSellerKycDraftRequest": {
        "type": "object",
        "required": ["account_type"],
        "properties": {
          "account_type": {
            "type": "string",
            "enum": ["individual", "business"]
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone_number": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string",
            "example": "Nigeria"
          },
          "bank_name": {
            "type": "string"
          },
          "bank_account_number": {
            "type": "string",
            "example": "0123456789"
          },
          "bank_account_name": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "date_of_birth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string"
          },
          "id_type": {
            "type": "string",
            "enum": [
              "national_id",
              "international_passport",
              "drivers_license",
              "voters_card"
            ]
          },
          "id_number": {
            "type": "string"
          },
          "business_name": {
            "type": "string"
          },
          "registration_number": {
            "type": "string"
          },
          "registered_business_address": {
            "type": "string"
          },
          "representative_first_name": {
            "type": "string"
          },
          "representative_last_name": {
            "type": "string"
          },
          "representative_role": {
            "type": "string"
          }
        }
      },
      "UploadSellerKycDocumentRequest": {
        "type": "object",
        "required": ["document_type", "file_name", "mime_type", "file_base64"],
        "properties": {
          "document_type": {
            "type": "string",
            "enum": [
              "proof_of_address",
              "id_document",
              "selfie",
              "cac_certificate"
            ]
          },
          "file_name": {
            "type": "string",
            "example": "utility-bill.jpg"
          },
          "mime_type": {
            "type": "string",
            "enum": ["image/jpeg", "image/png", "application/pdf"]
          },
          "file_base64": {
            "type": "string",
            "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQ..."
          }
        }
      },
      "SubmitSellerKycRequest": {
        "type": "object",
        "properties": {
          "confirm": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  }
}